pyproject.toml
[tool.poetry]
name = "aiogram-bot-template"
version = "0.1.0"
description = ""
authors = ["rkjar <rkjar@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.25.1"
python-dotenv = "^1.0.0"
redis = "^4.5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
